feat(score): add score system to the app
The hoop now has borders at the sides, to make sure the ball doesn't pass through the hoop's collider. The hoop has a child object, and when the ball triggers it, it scores. Also, a new SampleScene is added, to test the ball's collider and make sure it fits through the hoop, triggering the new object.